pyproject.toml
[project]
name = "mila-experiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "hydra-core>=1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "numpy>=2.0.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "wandb>=0.24.2",
]
